+++ jetstream/plugins/allocation_source.py
@@ -1,4 +1,4 @@
-from django.conf import settings
+from constance import config as constance_config
 
 from core.models import AllocationSource
 
@@ -44,9 +44,9 @@ def _get_enforcement_override(allocation_source):
         """
     assert isinstance(allocation_source, AllocationSource)
     import core.plugins
-    if allocation_source.name in getattr(settings, 'ALLOCATION_OVERRIDES_NEVER_ENFORCE', []):
+    if allocation_source.name in getattr(constance_config, 'ALLOCATION_OVERRIDES_NEVER_ENFORCE', '').split(','):
         return core.plugins.EnforcementOverrideChoice.NEVER_ENFORCE
-    if allocation_source.name in getattr(settings, 'ALLOCATION_OVERRIDES_ALWAYS_ENFORCE', []):
+    if allocation_source.name in getattr(constance_config, 'ALLOCATION_OVERRIDES_ALWAYS_ENFORCE', '').split(','):
         return core.plugins.EnforcementOverrideChoice.ALWAYS_ENFORCE
     return core.plugins.EnforcementOverrideChoice.NO_OVERRIDE